# swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "NestJS DDD template using nestia + prisma"
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "NestDDD - NestJS DDD template"
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "페이지 정보 1이상의 정수, default 1",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "상품 목록",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Invalid"
                    },
                    {
                      "$ref": "#/components/schemas/ISuccess_lt_PaginatedResponse_lt_IProduct.Summary_gt__gt_"
                    }
                  ]
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "상품 목록 조회 API",
        "description": "전체 상품 목록 조회",
        "x-nestia-namespace": "products.findMany",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "상품 목록 조회 API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "tag",
            "text": [
              {
                "text": "products",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "page",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "페이지 정보 1이상의 정수, default 1",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns",
            "text": [
              {
                "text": "상품 목록",
                "kind": "text"
              }
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "products"
        ],
        "parameters": [],
        "requestBody": {
          "description": "상품 생성 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IProduct.CreateInput"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "생성한 상품 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Invalid"
                    },
                    {
                      "$ref": "#/components/schemas/ISuccess_lt_IProduct.Detail_gt_"
                    },
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Forbidden"
                    }
                  ]
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "상품 생성 요청 API",
        "description": "",
        "x-nestia-namespace": "products.create",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "상품 생성 요청 API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "tag",
            "text": [
              {
                "text": "products",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "body",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "상품 생성 정보",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns",
            "text": [
              {
                "text": "생성한 상품 상세 정보",
                "kind": "text"
              }
            ]
          }
        ]
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": [
          "products"
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "description": "대상 상품 고유 번호",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "상품 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Invalid"
                    },
                    {
                      "$ref": "#/components/schemas/IFailure.Business.NotFound"
                    },
                    {
                      "$ref": "#/components/schemas/ISuccess_lt_IProduct.Detail_gt_"
                    }
                  ]
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "상품 상세 조회 API",
        "description": "",
        "x-nestia-namespace": "products.findOne",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "상품 상세 조회 API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "tag",
            "text": [
              {
                "text": "products",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "product_id",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "대상 상품 고유 번호",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns",
            "text": [
              {
                "text": "상품 상세 정보",
                "kind": "text"
              }
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "products"
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "description": "대상 상품 고유 번호",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "변경할 상품 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IProduct.UpdateInput"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "변경된 상품 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Invalid"
                    },
                    {
                      "$ref": "#/components/schemas/IFailure.Business.NotFound"
                    },
                    {
                      "$ref": "#/components/schemas/ISuccess_lt_IProduct.Detail_gt_"
                    },
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Forbidden"
                    }
                  ]
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "상품 수정 요청 API",
        "description": "",
        "x-nestia-namespace": "products.update",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "상품 수정 요청 API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "tag",
            "text": [
              {
                "text": "products",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "product_id",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "대상 상품 고유 번호",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "body",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "변경할 상품 정보",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns",
            "text": [
              {
                "text": "변경된 상품 상세 정보",
                "kind": "text"
              }
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "products"
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "description": "대상 상품 고유 번호",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "true",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Invalid"
                    },
                    {
                      "$ref": "#/components/schemas/IFailure.Business.NotFound"
                    },
                    {
                      "$ref": "#/components/schemas/IFailure.Business.Forbidden"
                    },
                    {
                      "$ref": "#/components/schemas/ISuccess_lt_true_gt_"
                    }
                  ]
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "상품 삭제(비활성화) 요청 API",
        "description": "",
        "x-nestia-namespace": "products.inActivate",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "상품 삭제(비활성화) 요청 API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "tag",
            "text": [
              {
                "text": "products",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "product_id",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "대상 상품 고유 번호",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns",
            "text": [
              {
                "text": "true",
                "kind": "text"
              }
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IFailure.Business.Invalid": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "business"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "event": {
            "type": "string",
            "enum": [
              "Invalid"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "message": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "type",
          "event",
          "message"
        ],
        "x-typia-jsDocTags": []
      },
...
      "ISuccess_lt_PaginatedResponse_lt_IProduct.Summary_gt__gt_": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "data": {
            "$ref": "#/components/schemas/PaginatedResponse_lt_IProduct.Summary_gt_",
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "type",
          "data"
        ],
        "x-typia-jsDocTags": []
      },
      "PaginatedResponse_lt_IProduct.Summary_gt_": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IProduct.Summary",
              "x-typia-required": true
            },
            "nullable": false,
            "x-typia-required": true
          },
          "page": {
            "type": "integer",
            "nullable": false,
            "x-typia-metaTags": [
              {
                "kind": "type",
                "value": "uint"
              },
              {
                "kind": "minimum",
                "value": 1
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "type",
                "text": [
                  {
                    "text": "uint",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minimum",
                "text": [
                  {
                    "text": "1",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "minimum": 1
          },
          "total_count": {
            "type": "number",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "data",
          "page",
          "total_count"
        ],
        "x-typia-jsDocTags": []
      },
      "IProduct.Summary": {
        "type": "object",
        "properties": {
          "vender": {
            "$ref": "#/components/schemas/IProduct.Vender",
            "x-typia-required": true
          },
          "name": {
            "type": "string",
            "nullable": false,
            "description": "상품명",
            "x-typia-required": true
          },
          "price": {
            "type": "integer",
            "nullable": false,
            "description": "상품 가격(원)",
            "x-typia-metaTags": [
              {
                "kind": "type",
                "value": "uint"
              },
              {
                "kind": "minimum",
                "value": 0
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "type",
                "text": [
                  {
                    "text": "uint",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minimum",
                "text": [
                  {
                    "text": "0",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "minimum": 0
          },
          "description": {
            "type": "string",
            "nullable": false,
            "description": "상품 설명",
            "x-typia-required": true
          },
          "id": {
            "type": "string",
            "nullable": false,
            "description": "상품 고유 id",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^[A-Z0-9]{10}$"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^[A-Z0-9]{10}$",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "pattern": "^[A-Z0-9]{10}$"
          },
          "created_at": {
            "type": "string",
            "nullable": false,
            "description": "상품 정보 생성 일시(ISO 8601 형식)",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "pattern": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$"
          }
        },
        "nullable": false,
        "required": [
          "vender",
          "name",
          "price",
          "description",
          "id",
          "created_at"
        ],
        "x-typia-jsDocTags": []
      },
      "IProduct.Vender": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          },
          "name": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "id",
          "name"
        ],
        "x-typia-jsDocTags": []
      },
      "IFailure.Business.NotFound": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "business"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "event": {
            "type": "string",
            "enum": [
              "NotFound"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "message": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "type",
          "event",
          "message"
        ],
        "x-typia-jsDocTags": []
      },
      "ISuccess_lt_IProduct.Detail_gt_": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "data": {
            "$ref": "#/components/schemas/IProduct.Detail",
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "type",
          "data"
        ],
        "x-typia-jsDocTags": []
      },
      "IProduct.Detail": {
        "type": "object",
        "properties": {
          "vender": {
            "$ref": "#/components/schemas/IProduct.Vender",
            "x-typia-required": true
          },
          "name": {
            "type": "string",
            "nullable": false,
            "description": "상품명",
            "x-typia-required": true
          },
          "price": {
            "type": "integer",
            "nullable": false,
            "description": "상품 가격(원)",
            "x-typia-metaTags": [
              {
                "kind": "type",
                "value": "uint"
              },
              {
                "kind": "minimum",
                "value": 0
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "type",
                "text": [
                  {
                    "text": "uint",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minimum",
                "text": [
                  {
                    "text": "0",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "minimum": 0
          },
          "description": {
            "type": "string",
            "nullable": false,
            "description": "상품 설명",
            "x-typia-required": true
          },
          "id": {
            "type": "string",
            "nullable": false,
            "description": "상품 고유 id",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^[A-Z0-9]{10}$"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^[A-Z0-9]{10}$",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "pattern": "^[A-Z0-9]{10}$"
          },
          "created_at": {
            "type": "string",
            "nullable": false,
            "description": "상품 정보 생성 일시(ISO 8601 형식)",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "pattern": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$"
          },
          "is_deleted": {
            "type": "boolean",
            "nullable": false,
            "description": "구매 가능 상태",
            "x-typia-required": true
          },
          "updated_at": {
            "type": "string",
            "nullable": false,
            "description": "상품 정보 최근 수정 일시(ISO 8601 형식)",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "pattern": "^\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]).[0-9]{3}Z$"
          }
        },
        "nullable": false,
        "required": [
          "vender",
          "name",
          "price",
          "description",
          "id",
          "created_at",
          "is_deleted",
          "updated_at"
        ],
        "x-typia-jsDocTags": []
      },
      "IProduct.CreateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": false,
            "description": "상품명",
            "x-typia-required": true
          },
          "price": {
            "type": "integer",
            "nullable": false,
            "description": "상품 가격(원)",
            "x-typia-metaTags": [
              {
                "kind": "type",
                "value": "uint"
              },
              {
                "kind": "minimum",
                "value": 0
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "type",
                "text": [
                  {
                    "text": "uint",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minimum",
                "text": [
                  {
                    "text": "0",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "minimum": 0
          },
          "description": {
            "type": "string",
            "nullable": false,
            "description": "상품 설명",
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "name",
          "price",
          "description"
        ],
        "x-typia-jsDocTags": []
      },
      "IFailure.Business.Forbidden": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "business"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "event": {
            "type": "string",
            "enum": [
              "Forbidden"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "message": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "type",
          "event",
          "message"
        ],
        "x-typia-jsDocTags": []
      },
      "IProduct.UpdateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": false,
            "description": "상품명",
            "x-typia-required": false
          },
          "price": {
            "type": "integer",
            "nullable": false,
            "description": "상품 가격(원)",
            "x-typia-metaTags": [
              {
                "kind": "type",
                "value": "uint"
              },
              {
                "kind": "minimum",
                "value": 0
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "type",
                "text": [
                  {
                    "text": "uint",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minimum",
                "text": [
                  {
                    "text": "0",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": false,
            "minimum": 0
          },
          "description": {
            "type": "string",
            "nullable": false,
            "description": "상품 설명",
            "x-typia-required": false
          }
        },
        "nullable": false,
        "x-typia-jsDocTags": []
      },
      "ISuccess_lt_true_gt_": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false,
            "x-typia-required": true
          },
          "data": {
            "type": "boolean",
            "enum": [
              true
            ],
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "type",
          "data"
        ],
        "x-typia-jsDocTags": []
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}